apps/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic_settings
pydantic
asyncpg
google-genai

# Deprecated Soon
google-cloud-aiplatform